Move credentials.js login.js to app/session/ dir
